Abstract out stuff. too much for one commit. c'est la vie.
